Remove substrate-bip39 dependency
tiny-bip39 offers a Seed::new which seems completely equivalent *in purpose* to
the single used function from substrate-bip39. Accordingly, substrate-bip39
appears unneeded.

It technically isn't equivalent as substrate-bip39 runs pbkdf2 on the entropy,
when bip39 specifies to do so on the mnemonic (as tiny-bip39 does). Vectors
have been updated accordingly.

The alternative would be to merge substrate-bip39 into the monorepo, yet that
can't be done as our substrate-bip39 is on dalek 4 and this repo is on dalek 3
prior to the next libp2p update being merged into the upstream.

This also reduces our dependency tree (and technically, our spec count).