Finish rebasing
Corrects errors made in the process and updates newly introduced components to
libp2p 0.52.

Edited due to further rebase onto 1.5.0. This rebase additional reverted
effectively all network changes re: notification_service unfortunately, due to
the incompatibilies with the pending libp2p PR. This is accepted as this is an
experimental branch explicitly not to be actually used.