Modify update_storage to accept the root effected
All call sites *have* the root. The in_mem backend would need to recalculate it
itself to avoid fetching it from the header, if not for this.

The in_mem backend appears only used for tests. That means we're modifying a
trait for tests, while unused in prod. I have mixed feelings on that.